Refactor: 모의면접 관련 Realm 클래스도 통합
